requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
dotenv==0.9.9
drf-spectacular==0.27.2
filelock==3.25.2
firebase_admin==7.4.0
fsspec==2026.3.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.78.0
grpcio-status==1.62.3
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.23
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lxml==6.0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.3
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pluggy==1.6.0
portalocker==3.2.0
proto-plus==1.27.2
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.10.1
pypdf==6.10.2
pytest==9.0.3
python-docx==1.2.0
python-dotenv==1.2.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.1
referencing==0.37.0
regex==2026.3.32
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.5
stripe==15.0.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
tiktoken
sentence-transformers
pypdf
python-docx
langgraph